refactor: move confirmation dialog to ui directory
