+++ src/components/Agents/DeleteAgentButton.tsx
@@ -1,8 +1,8 @@
 import { useCallback, useState } from "react";
 import { FaCircleNotch, FaTrash } from "react-icons/fa";
 import { useDeleteAgentMutations } from "../../api/query-hooks/mutations/useUpsertAgentMutations";
+import { ConfirmationPromptDialog } from "../../ui/AlertDialog/ConfirmationPromptDialog";
 import { Button } from "../../ui/Button";
-import { ConfirmationPromptDialog } from "../Dialogs/ConfirmationPromptDialog";
 import { toastError, toastSuccess } from "../Toast/toast";
 
 type DeleteAgentButtonProps = {
