+++ src/components/SchemaResourcePage/Delete/DeleteResource.tsx
@@ -1,8 +1,8 @@
 import { useSettingsDeleteResource } from "@flanksource-ui/api/query-hooks/mutations/useSettingsResourcesMutations";
+import { ConfirmationPromptDialog } from "@flanksource-ui/ui/AlertDialog/ConfirmationPromptDialog";
 import { Button } from "@flanksource-ui/ui/Button";
 import { useCallback, useState } from "react";
 import { FaCircleNotch, FaTrash } from "react-icons/fa";
-import { ConfirmationPromptDialog } from "../../Dialogs/ConfirmationPromptDialog";
 import { toastSuccess } from "../../Toast/toast";
 import { SchemaResourceType } from "../resourceTypes";
